lobby button is disabled till connection is established, code clean up, second castle, Cannon.pointAt returns true if the rotation has been changed and code cleanup stuff...

+++ src/com/angrykings/activities/LocalTestActivity.java
@@ -311,7 +311,7 @@ public class LocalTestActivity extends BaseGameActivity
 			float x = pSceneTouchEvent.getX();
 			float y = pSceneTouchEvent.getY();
 
-			this.cannon.pointAt(x, y);
+			this.cannon.pointAt((int)x, (int)y);
 
 			if (pSceneTouchEvent.isActionUp()) {
 				this.cannon.fire(400);
